% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimindex1.R
\name{kimindex1}
\alias{kimindex1}
\title{Calculate Grain Area from Grain Weight (Kim Index1) of Bread Wheat}
\usage{
kimindex1(data, weight_col, remove_na = FALSE)
}
\arguments{
\item{data}{A data frame containing the grain weight}

\item{weight_col}{The name of the column in the data frame that contains the grain weight. This name can be any column that holds the grain weight data.}
}
\value{
The input data frame with an additional column \code{predicted_area} representing the predicted individual grain area from each grain weight.
}
\description{
This function predicts the grain area (mm2) for wheat based on the grain weight (mg)
using the equation y = exp(log(x)/1.32), where y is the grain area and x is the grain weight.
}
\examples{
predicted_grain_area= kim_index1(data, "grain_weight", remove_na= FALSE OR TRUE)

}
